+++ Plutarch/Internal/TermCont.hs
@@ -6,7 +6,6 @@ module Plutarch.Internal.TermCont (
   runTermCont,
   unTermCont,
   tcont,
-  pfindPlaceholder,
   pfindAllPlaceholders,
 ) where
 
@@ -20,6 +19,7 @@ import Plutarch.Internal.Term (
   RawTerm (..),
   S,
   Term (Term),
+  TermResult (TermResult),
   TracingMode (DetTracing),
   asRawTerm,
   getTerm,
@@ -62,57 +62,35 @@ tcont :: ((a -> Term s r) -> Term s r) -> TermCont @r s a
 tcont = TermCont
 
 hashOpenTerm :: Term s a -> TermCont s (Hashed RawTerm)
-hashOpenTerm x = TermCont $ \f -> Term $ \i -> do
-  y <- asRawTerm x i
-  let h = hashed $ getTerm y
-  asRawTerm (f h) i
+hashOpenTerm x = TermCont $ \f -> Term $ do
+  TermResult t _ <- asRawTerm x
+  let h = hashed t
+  asRawTerm (f h)
 
 -- This can technically be done outside of TermCont.
 -- Need to pay close attention when killing branch with this.
 -- If term is pre-evaluated (via `evalTerm`), RawTerm will no longer hold
 -- tagged RPlaceholder.
 
-{- | Given a term, and an integer tag, this function checks if the term holds and
-@PPlaceholder@ with the given integer tag.
--}
-pfindPlaceholder :: Integer -> Term s a -> TermCont s Bool
-pfindPlaceholder idx x = TermCont $ \f -> Term $ \i -> do
-  y <- asRawTerm x i
-  asRawTerm (f . findPlaceholder . getTerm $ y) i
-  where
-    findPlaceholder = \case
-      RLamAbs _ x -> findPlaceholder x
-      RApply x xs -> any findPlaceholder (x : xs)
-      RForce x -> findPlaceholder x
-      RDelay x -> findPlaceholder x
-      RHoisted (HoistedTerm _ x) -> findPlaceholder x
-      RPlaceHolder idx' -> idx == idx'
-      RConstr _ xs -> any findPlaceholder xs
-      RCase x xs -> any findPlaceholder (x : xs)
-      RVar _ -> False
-      RConstant _ -> False
-      RBuiltin _ -> False
-      RCompiled _ -> False
-      RError -> False
-
 -- | Finds all placeholder ids and returns it
 pfindAllPlaceholders :: Term s a -> TermCont s [Integer]
-pfindAllPlaceholders x = TermCont $ \f -> Term $ \i -> do
-  y <- asRawTerm x i
-  asRawTerm (f . nub . findPlaceholder . getTerm $ y) i
+pfindAllPlaceholders x = TermCont $ \f -> Term $ do
+  y@(TermResult yRaw _) <- asRawTerm x
+  asRawTerm (f . nub . findPlaceholder yRaw . getTerm $ y)
   where
-    findPlaceholder :: RawTerm -> [Integer]
-    findPlaceholder = \case
-      RLamAbs _ x -> findPlaceholder x
-      RApply x xs -> findPlaceholder x <> foldMap findPlaceholder xs
-      RForce x -> findPlaceholder x
-      RDelay x -> findPlaceholder x
-      RHoisted (HoistedTerm _ x) -> findPlaceholder x
+    findPlaceholder :: RawTerm -> RawTerm -> [Integer]
+    findPlaceholder t = \case
+      RLamAbs _ x -> findPlaceholder t x
+      RApply x xs -> findPlaceholder t x <> foldMap (findPlaceholder t) xs
+      RForce x -> findPlaceholder t x
+      RDelay x -> findPlaceholder t x
+      RHoisted (HoistedTerm _ x) -> findPlaceholder t x
       RPlaceHolder idx -> [idx]
-      RConstr _ xs -> foldMap findPlaceholder xs
-      RCase x xs -> findPlaceholder x <> foldMap findPlaceholder xs
+      RConstr _ xs -> foldMap (findPlaceholder t) xs
+      RCase x xs -> findPlaceholder t x <> foldMap (findPlaceholder t) xs
       RVar _ -> []
       RConstant _ -> []
       RBuiltin _ -> []
       RCompiled _ -> []
       RError -> []
+      RLet v f -> findPlaceholder t v <> findPlaceholder t f