+++ plutarch-ledger-api/CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+## 3.7.0 -- 26/05/2026
+
+### Added
+
+* `Plutarch.LedgerApi.V3.Value` module, containing compatibility functionality
+  between various @Value@ representations and 'PBuiltinValue'
+* `PLiftable` instance for `PAsData PMintValue` by way of `MintValue`
+
+### Removed
+
+* `passertSorted` (both versions), due to confusing naming. Use
+  `ppromoteToSorted*` functions instead
+
 ## 3.6.0 -- 04/03/2025 
 
 ### Added