+++ plutarch-orphanage/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## UNRELEASED
+
+## 1.3.0 -- 26-05-2026
+
+### Added
+
+* `Arbitrary`, `CoArbitrary` and `Function` instances for types from
+  `PlutusCore.Value`.
+
 ## 1.2.0 -- 21-08-2025
 
 ### Added